requirements.txt
aiosqlite
aiofiles
nonebot_plugin_htmlrender